--- apps/explorer-e2e/src/support/app.po.ts
@@ -1,1 +1,0 @@
-export const getGreeting = () => cy.get('h1');
