--- apps/explorer-e2e/src/support/commands.ts
@@ -1,57 +1,0 @@
-// ***********************************************
-// This example commands.js shows you how to
-// create various custom commands and overwrite
-// existing commands.
-//
-// For more comprehensive examples of custom
-// commands please read more here:
-// https://on.cypress.io/custom-commands
-// ***********************************************
-
-// eslint-disable-next-line @typescript-eslint/no-namespace
-declare namespace Cypress {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  interface Chainable<Subject> {
-    login(email: string, password: string): void;
-    getByTestId(selector: string): Chainable<JQuery<HTMLElement>>;
-    slack(message: string): void;
-  }
-}
-//
-// -- This is a parent command --
-Cypress.Commands.add('login', (email, password) => {
-  console.log('Custom command example: Login', email, password);
-});
-//
-// -- This is a child command --
-// Cypress.Commands.add("drag", { prevSubject: 'element'}, (subject, options) => { ... })
-//
-//
-// -- This is a dual command --
-// Cypress.Commands.add("dismiss", { prevSubject: 'optional'}, (subject, options) => { ... })
-//
-//
-// -- This will overwrite an existing command --
-// Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
-Cypress.Commands.add('getByTestId', (selector, ...args) => {
-  return cy.get(`[data-testid=${selector}]`, ...args);
-});
-
-/**
- * Posts a message to slack #ui-notify
- */
-Cypress.Commands.add('slack', (message) => {
-  const text = `${message}: ${JSON.stringify(Cypress.spec)}`;
-
-  cy.log('NOTIFYING SLACK');
-
-  const webhook = Cypress.env('slackWebhook');
-
-  if (!webhook) {
-    return;
-  }
-
-  cy.request('POST', webhook, {
-    text,
-  });
-});
