+++ apps/stats-e2e/src/support/index.ts
@@ -13,5 +13,4 @@
 // https://on.cypress.io/configuration
 // ***********************************************************
 
-// Import commands.js using ES2015 syntax:
-import './commands';
+import '@vegaprotocol/cypress';
