+++ libs/cypress/README.md
@@ -1,0 +1,7 @@
+# cypress
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test cypress` to execute the unit tests via [Jest](https://jestjs.io).
