+++ libs/cypress/project.json
@@ -1,0 +1,23 @@
+{
+  "root": "libs/cypress",
+  "sourceRoot": "libs/cypress/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/cypress/**/*.ts"]
+      }
+    },
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["coverage/libs/cypress"],
+      "options": {
+        "jestConfig": "libs/cypress/jest.config.js",
+        "passWithNoTests": true
+      }
+    }
+  },
+  "tags": []
+}
