+++ libs/cypress/src/index.ts
@@ -1,0 +1,5 @@
+import { addGetTestIdcommand } from './lib/commands/get-by-test-id';
+import { addSlackCommand } from './lib/commands/slack';
+
+addGetTestIdcommand();
+addSlackCommand();
