+++ libs/cypress/src/lib/commands/get-by-test-id.ts
@@ -1,0 +1,14 @@
+// eslint-disable-next-line @typescript-eslint/no-namespace
+declare namespace Cypress {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  interface Chainable<Subject> {
+    getByTestId(selector: string): Chainable<JQuery<HTMLElement>>;
+  }
+}
+
+export function addGetTestIdcommand() {
+  // @ts-ignore - ignoring Cypress type error which gets resolved when Cypress uses the command
+  Cypress.Commands.add('getByTestId', (selector, ...args) => {
+    return cy.get(`[data-testid=${selector}]`, ...args);
+  });
+}
